Graphics.UI.Threepenny.Attributes

Synopsis

Element attributes.

Input elements

checked :: Attr Element Bool

The checked status of an input element of type checkbox.

selection :: Attr Element (Maybe Int)

Index of the currently selected option of a select element.

The index starts at 0. If no option is selected, then the selection is Nothing.

enabled :: Attr Element Bool

The enabled status of an input element

HTML attributes

action :: WriteAttr Element String

align :: WriteAttr Element String

alink :: WriteAttr Element String

alt :: WriteAttr Element String

altcode :: WriteAttr Element String

archive :: WriteAttr Element String

background :: WriteAttr Element String

base :: WriteAttr Element String

bgcolor :: WriteAttr Element String

border :: WriteAttr Element Int

bordercolor :: WriteAttr Element String

cellpadding :: WriteAttr Element Int

cellspacing :: WriteAttr Element Int

checked_ :: WriteAttr Element Bool

class_ :: WriteAttr Element String

clear_ :: WriteAttr Element String

code_ :: WriteAttr Element String

codebase :: WriteAttr Element String

color :: WriteAttr Element String

cols :: WriteAttr Element String

colspan :: WriteAttr Element Int

compact :: WriteAttr Element Bool

content :: WriteAttr Element String

coords :: WriteAttr Element String

enctype :: WriteAttr Element String

face :: WriteAttr Element String

for :: WriteAttr Element String

frameborder :: WriteAttr Element Int

height :: WriteAttr Element Int

href :: WriteAttr Element String

hspace :: WriteAttr Element Int

httpequiv :: WriteAttr Element String

id_ :: WriteAttr Element String

ismap :: WriteAttr Element Bool

lang :: WriteAttr Element String

marginheight :: WriteAttr Element Int

marginwidth :: WriteAttr Element Int

maxlength :: WriteAttr Element Int

method :: WriteAttr Element String

multiple :: WriteAttr Element Bool

name :: WriteAttr Element String

nohref :: WriteAttr Element Bool

noresize :: WriteAttr Element Bool

noshade :: WriteAttr Element Bool

nowrap :: WriteAttr Element Bool

rel :: WriteAttr Element String

rev :: WriteAttr Element String

rows :: WriteAttr Element String

rowspan :: WriteAttr Element Int

rules :: WriteAttr Element String

scrolling :: WriteAttr Element String

selected :: WriteAttr Element Bool

shape :: WriteAttr Element String

size :: WriteAttr Element String

src :: WriteAttr Element String

target :: WriteAttr Element String

text_ :: WriteAttr Element String

title__ :: WriteAttr Element String

type_ :: WriteAttr Element String

usemap :: WriteAttr Element String

valign :: WriteAttr Element String

version :: WriteAttr Element String

vlink :: WriteAttr Element String

vspace :: WriteAttr Element Int

width :: WriteAttr Element Int