Graphics.UI.Threepenny.Core

Synopsis

Core functionality of the Threepenny GUI library.

Server

To display the user interface, you have to start a server using startGUI. Then, visit the URL http://localhost:8023/ in your browser (assuming that you use the default server configuration defaultConfig, or have set the port number to jsPort=Just 8023.)

The server is multithreaded. FFI calls can be made concurrently, but events are handled sequentially.

FFI calls can be buffered, so in some circumstances, it may happen that you manipulate the browser window, but the effect is not immediately visible. See CallBufferMode for more information.

data Config

Static configuration for a Foreign.JavaScript server.

This is a record type which has the following fields:

jsPort :: Maybe Int

Port number. Nothing means that the port number is read from the environment variable PORT. Alternatively, port 8023 is used if this variable is not set.

jsAddr :: Maybe ByteString

Bind address. Nothing means that the bind address is read from the environment variable ADDR. Alternatively, address 127.0.0.1 is used if this variable is not set.

jsCustomHTML :: Maybe FilePath

Custom HTML file to replace the default one.

jsStatic :: Maybe FilePath

Directory that is served under /static.

jsLog :: ByteString -> IO ()

Function to print a single log message.

jsWindowReloadOnDisconnect :: Bool

Reload the browser window if the connection to the server was dropped accidentally, for instance because the computer was put to sleep and awoken again.

jsCallBufferMode :: CallBufferMode

The initial CallBufferMode to use for runFunction. It can be changed at any time with setCallBufferMode.

jsUseSSLBind :: Maybe ConfigSSL

Whether to serve on a HTTPS connection instead of HTTP for improved security.

Just with a ConfigSSL to serve on HTTPS. Note that this will fail silently unless the snap-server package has been compiled with the openssl flag enabled.

Nothing to serve on HTTP.

(For reasons of forward compatibility, the constructor is not exported.)

data ConfigSSL

Static configuration for the SSL version of the Foreign.JavaScript server.

This is a record type which has the following fields:

jsSSLBind :: ByteString

Bind address.

jsSSLCert :: FilePath

Path to SSL certificate file. Example: cert.pem.

jsSSLChainCert :: Bool

If it is SSL chain certificate file.

jsSSLKey :: FilePath

Path to SSL key file. Example: key.pem.

jsSSLPort :: ByteString

Port number. Example: 443.

Constructors

ConfigSSL

Fields

jsSSLBind :: ByteString

jsSSLCert :: FilePath

jsSSLChainCert :: Bool

jsSSLKey :: FilePath

jsSSLPort :: Int

defaultConfig :: Config

Default configuration.

Port from environment variable or 8023, listening on localhost, no custom HTML, no static directory, logging to stderr, do reload on disconnect, buffer FFI calls.

startGUI

Arguments

:: Config
Server configuration.

-> (Window -> UI ())
Action to run whenever a client browser connects.

-> IO ()

Start server for GUI sessions.

loadFile

Arguments

:: String
MIME type

-> FilePath
Local path to the file

-> UI String
Relative URI under which this file is now accessible

Begin to serve a local file with a given MimeType under a relative URI.

loadDirectory :: FilePath -> UI String

Make a local directory available under a relative URI.

UI monad

data UI a

User interface elements are created and manipulated in the UI monad.

This monad is essentially just a thin wrapper around the familiar IO monad. Use the liftIO function to access IO operations like reading and writing from files.

There are several subtle reasons why Threepenny uses a custom UI monad instead of the standard IO monad:

More convenience when calling JavaScript. The monad keeps track of a browser Window context in which JavaScript function calls are executed.

Recursion for functional reactive programming.

runUI :: Window -> UI a -> IO a

Execute an UI action in a particular browser window. Also runs all scheduled IO actions.

class Monad m => MonadUI m where

Methods

liftUI :: UI a -> m a

Lift a computation from the UI monad.

askWindow :: UI Window

Retrieve current Window context in the UI monad.

liftIOLater :: IO () -> UI ()

Schedule an IO action to be run later.

module Control.Monad.IO.Class

module Control.Monad.Fix

Browser Window

data Window

The type Window represents a browser window.

title :: WriteAttr Window String

Title of the client window.

DOM elements

Create and manipulate DOM elements.

data Element

getWindow :: Element -> IO Window

mkElement :: String -> UI Element

Make a new DOM element with a given tag name.

mkElementNamespace :: Maybe String -> String -> UI Element

Make a new DOM element with a namespace and a given tag name.

A namespace Nothing corresponds to the default HTML namespace.

delete :: Element -> UI ()

Delete the given element.

This operation removes the element from the browser window DOM and marks it for garbage collection on the Haskell side. The element is unusable afterwards.

NOTE: If you wish to temporarily remove an element from the DOM tree, change the children property of its parent element instead.

string :: String -> UI Element

Make a span element with a given text content.

getHead :: Window -> UI Element

Get the head of the page.

getBody :: Window -> UI Element

Get the body of the page.

(#+) :: UI Element -> [UI Element] -> UI Element infixl 8

Append DOM elements as children to a given element.

children :: WriteAttr Element [Element]

Child elements of a given element.

text :: WriteAttr Element String

Text content of an element.

html :: WriteAttr Element String

Child elements of a given element as a HTML string.

attr :: String -> WriteAttr Element String

HTML attributes of an element.

style :: WriteAttr Element [(String, String)]

Set CSS style of an Element

value :: Attr Element String

Value attribute of an element. Particularly relevant for control widgets like input.

getElementsByTagName

Arguments

:: Window
Browser window

-> String
The tag name.

-> UI [Element]
All elements with that tag name.

Get all elements of the given tag name.

getElementById

Arguments

:: Window
Browser window

-> String
The ID string.

-> UI (Maybe Element)
Element (if any) with given ID.

Get an element by a particular ID.

getElementsByClassName

Arguments

:: Window
Browser window

-> String
The class string.

-> UI [Element]
Elements with given class.

Get a list of elements by particular class.

Layout

Combinators for quickly creating layouts. They can be adjusted with CSS later on.

grid :: [[UI Element]] -> UI Element

Align given elements in a rectangular grid.

Layout is achieved by using the CSS display:table property. The following element tree will be generated

 <div class="table">
   <div class="table-row">
     <div class="table-cell"> ... </div>
     <div class="table-cell"> ... </div>
   </div>
   <div class="table-row">
     ...
   </div>
  ...
  </div>

You can customatize the actual layout by assigning an id to the element and changing the .table, .table-row and table-column classes in a custom CSS file.

row :: [UI Element] -> UI Element

Align given elements in a row. Special case of grid.

column :: [UI Element] -> UI Element

Align given elements in a column. Special case of grid.

Events

For a list of predefined events, see Graphics.UI.Threepenny.Events.

type EventData = Value

Events may carry data. At the moment, they may return a single JSON value, as defined in the Data.Aeson module.

domEvent

Arguments

:: String
Event name. A full list can be found at http://www.w3schools.com/jsref/dom_obj_event.asp. Note that the on-prefix is not included, the name is click and so on.

-> Element
Element where the event is to occur.

-> Event EventData

Obtain DOM event for a given element.

unsafeFromJSON :: FromJSON a => EventData -> a

Convert event data to a Haskell value. Throws an exception when the data cannot be converted.

disconnect :: Window -> Event ()

Event that occurs whenever the client has disconnected, be it by closing the browser window or by exception.

Note: DOM Elements in a browser window that has been closed can no longer be manipulated.

on :: (element -> Event a) -> element -> (a -> UI void) -> UI ()

Convenience function to register Events for Elements.

Example usage.

on click element $ \_ -> ...

onEvent :: Event a -> (a -> UI void) -> UI (UI ())

Register an UI action to be executed whenever the Event happens.

FIXME: Should be unified with on?

onChanges :: Behavior a -> (a -> UI void) -> UI ()

Execute a UI action whenever a Behavior changes. Use sparingly, it is recommended that you use sink instead.

class Functor f => Applicative (f :: Type -> Type) where

A functor with application, providing operations to

embed pure expressions (pure), and

sequence computations and combine their results (<*> and liftA2).

A minimal complete definition must include implementations of pure and of either <*> or liftA2. If it defines both, then they must behave the same as their default definitions:

(<*>) = liftA2 id

liftA2 f x y = f <$> x <*> y

Further, any definition must satisfy the following:

Identity

pure id <*> v = v

Composition

pure (.) <*> u <*> v <*> w = u <*> (v <*> w)

Homomorphism

pure f <*> pure x = pure (f x)

Interchange

u <*> pure y = pure ($ y) <*> u

The other methods have the following default definitions, which may be overridden with equivalent specialized implementations:

u *> v = (id <$ u) <*> v

u <* v = liftA2 const u v

As a consequence of these laws, the Functor instance for f will satisfy

fmap f x = pure f <*> x

It may be useful to note that supposing

forall x y. p (q x y) = f x . g y

it follows from the above that

liftA2 p (liftA2 q u v) = liftA2 f u . liftA2 g v

If f is also a Monad, it should satisfy

pure = return

m1 <*> m2 = m1 >>= (\x1 -> m2 >>= (\x2 -> return (x1 x2)))

(*>) = (>>)

(which implies that pure and <*> satisfy the applicative functor laws).

Minimal complete definition

pure, ((<*>) | liftA2)

Methods

pure :: a -> f a

Lift a value.

(<*>) :: f (a -> b) -> f a -> f b infixl 4

Sequential application.

A few functors support an implementation of <*> that is more efficient than the default one.

Example

Expand

Used in combination with (<$>), (<*>) can be used to build a record.

>>> data MyState = MyState {arg1 :: Foo, arg2 :: Bar, arg3 :: Baz}

>>> produceFoo :: Applicative f => f Foo

>>> produceBar :: Applicative f => f Bar
>>> produceBaz :: Applicative f => f Baz

>>> mkState :: Applicative f => f MyState
>>> mkState = MyState <$> produceFoo <*> produceBar <*> produceBaz

liftA2 :: (a -> b -> c) -> f a -> f b -> f c

Lift a binary function to actions.

Some functors support an implementation of liftA2 that is more efficient than the default one. In particular, if fmap is an expensive operation, it is likely better to use liftA2 than to fmap over the structure and then use <*>.

This became a typeclass method in 4.10.0.0. Prior to that, it was a function defined in terms of <*> and fmap.

Example

Expand

>>> liftA2 (,) (Just 3) (Just 5)
Just (3,5)

(*>) :: f a -> f b -> f b infixl 4

Sequence actions, discarding the value of the first argument.

Examples

Expand

If used in conjunction with the Applicative instance for Maybe, you can chain Maybe computations, with a possible "early return" in case of Nothing.

>>> Just 2 *> Just 3
Just 3

>>> Nothing *> Just 3
Nothing

Of course a more interesting use case would be to have effectful computations instead of just returning pure values.

>>> import Data.Char
>>> import Text.ParserCombinators.ReadP
>>> let p = string "my name is " *> munch1 isAlpha <* eof
>>> readP_to_S p "my name is Simon"
[("Simon","")]

(<*) :: f a -> f b -> f a infixl 4

Sequence actions, discarding the value of the second argument.

class Applicative f => Alternative (f :: Type -> Type) where

A monoid on applicative functors.

If defined, some and many should be the least solutions of the equations:

some v = (:) <$> v <*> many v

many v = some v <|> pure []

Minimal complete definition

empty, (<|>)

Methods

empty :: f a

The identity of <|>

(<|>) :: f a -> f a -> f a infixl 3

An associative binary operation

some :: f a -> f [a]

One or more.

many :: f a -> f [a]

Zero or more.

newtype Const a (b :: k)

The Const functor.

Constructors

Const

Fields

getConst :: a

type Handler a = a -> IO ()

An event handler is a function that takes an event value and performs some computation.

data Event a

Event a represents a stream of events as they occur in time. Semantically, you can think of Event a as an infinite list of values that are tagged with their corresponding time of occurence,

type Event a = [(Time,a)]

newtype ZipList a

Lists, but with an Applicative functor based on zipping.

Constructors

ZipList

Fields

getZipList :: [a]

newtype WrappedArrow (a :: Type -> Type -> Type) b c

Constructors

WrapArrow

Fields

unwrapArrow :: a b c

newtype WrappedMonad (m :: Type -> Type) a

Constructors

WrapMonad

Fields

unwrapMonad :: m a

data Tidings a

Data type representing a behavior (facts) and suggestions to change it (rumors).

data Behavior a

Behavior a represents a value that varies in time. Think of it as

type Behavior a = Time -> a

(<$>) :: Functor f => (a -> b) -> f a -> f b infixl 4

An infix synonym for fmap.

The name of this operator is an allusion to $. Note the similarities between their types:

 ($)  ::              (a -> b) ->   a ->   b
(<$>) :: Functor f => (a -> b) -> f a -> f b

Whereas $ is function application, <$> is function application lifted over a Functor.

Examples

Expand

Convert from a Maybe Int to a Maybe String using show:

>>> show <$> Nothing
Nothing
>>> show <$> Just 3
Just "3"

Convert from an Either Int Int to an Either Int String using show:

>>> show <$> Left 17
Left 17
>>> show <$> Right 17
Right "17"

Double each element of a list:

>>> (*2) <$> [1,2,3]
[2,4,6]

Apply even to the second element of a pair:

>>> even <$> (2,2)
(2,True)

(<$) :: Functor f => a -> f b -> f a infixl 4

Replace all locations in the input with the same value. The default definition is fmap . const, but this may be overridden with a more efficient version.

Examples

Expand

Perform a computation with Maybe and replace the result with a constant value if it is Just:

>>> 'a' <$ Just 2
Just 'a'
>>> 'a' <$ Nothing
Nothing

(<**>) :: Applicative f => f a -> f (a -> b) -> f b infixl 4

A variant of <*> with the arguments reversed.

liftA :: Applicative f => (a -> b) -> f a -> f b

Lift a function to actions. Equivalent to Functor's fmap but implemented using only Applicative's methods: liftA f a = pure f <*> a

As such this function may be used to implement a Functor instance from an Applicative one.

Examples

Expand

Using the Applicative instance for Lists:

>>> liftA (+1) [1, 2]
[2,3]

Or the Applicative instance for Maybe

>>> liftA (+1) (Just 3)
Just 4

liftA3 :: Applicative f => (a -> b -> c -> d) -> f a -> f b -> f c -> f d

Lift a ternary function to actions.

optional :: Alternative f => f a -> f (Maybe a)

One or none.

It is useful for modelling any computation that is allowed to fail.

Examples

Expand

Using the Alternative instance of Control.Monad.Except, the following functions:

>>> import Control.Monad.Except

>>> canFail = throwError "it failed" :: Except String Int
>>> final = return 42                :: Except String Int

Can be combined by allowing the first function to fail:

>>> runExcept $ canFail *> final
Left "it failed"
>>> runExcept $ optional canFail *> final
Right 42

apply :: Behavior (a -> b) -> Event a -> Event b

Apply a time-varying function to a stream of events. Think of it as

apply bf ex = [(time, bf time x) | (time, x) <- ex]

asum :: (Foldable t, Alternative f) => t (f a) -> f a

The sum of a collection of actions using (<|>), generalizing concat.

asum is just like msum, but generalised to Alternative.

Examples

Expand

Basic usage:

>>> asum [Just "Hello", Nothing, Just "World"]
Just "Hello"

split :: Event (Either a b) -> (Event a, Event b)

Split event occurrences according to a tag. The Left values go into the left component while the Right values go into the right component of the result.

unions :: [Event a] -> Event [a]

Collect simultaneous event occurrences in a list.

unionWith :: (a -> a -> a) -> Event a -> Event a -> Event a

Merge two event streams of the same type. In case of simultaneous occurrences, the event values are combined with the binary function. Think of it as

unionWith f ((timex,x):xs) ((timey,y):ys)
   | timex == timey = (timex,f x y) : unionWith f xs ys
   | timex <  timey = (timex,x)     : unionWith f xs ((timey,y):ys)
   | timex >  timey = (timey,y)     : unionWith f ((timex,x):xs) ys

mapAccum :: MonadIO m => acc -> Event (acc -> (x, acc)) -> m (Event x, Behavior acc)

Efficient combination of accumE and accumB.

tidings :: Behavior a -> Event a -> Tidings a

Smart constructor. Combine facts and rumors into Tidings.

newEvent :: IO (Event a, Handler a)

Create a new event. Also returns a function that triggers an event occurrence.

register :: Event a -> Handler a -> IO (IO ())

Register an event Handler for an Event. All registered handlers will be called whenever the event occurs.

When registering an event handler, you will also be given an action that unregisters this handler again.

do unregisterMyHandler <- register event myHandler

FIXME: Unregistering event handlers does not work yet.

currentValue :: MonadIO m => Behavior a -> m a

Read the current value of a Behavior.

never :: Event a

Event that never occurs. Think of it as never = [].

filterJust :: Event (Maybe a) -> Event a

Return all event occurrences that are Just values, discard the rest. Think of it as

filterJust es = [(time,a) | (time,Just a) <- es]

accumE :: MonadIO m => a -> Event (a -> a) -> m (Event a)

The accumE function accumulates a stream of events. Example:

accumE "x" [(time1,(++"y")),(time2,(++"z"))]
   = return [(time1,"xy"),(time2,"xyz")]

Note that the output events are simultaneous with the input events, there is no "delay" like in the case of accumB.

stepper :: MonadIO m => a -> Event a -> m (Behavior a)

Construct a time-varying function from an initial value and a stream of new values. Think of it as

stepper x0 ex = return $ \time ->
    last (x0 : [x | (timex,x) <- ex, timex < time])

Note that the smaller-than-sign in the comparison timex < time means that the value of the behavior changes "slightly after" the event occurrences. This allows for recursive definitions.

(<@>) :: Behavior (a -> b) -> Event a -> Event b infixl 4

Infix synonym for apply, similar to <*>.

(<@) :: Behavior a -> Event b -> Event a infixl 4

Variant of apply similar to <*

filterE :: (a -> Bool) -> Event a -> Event a

Return all event occurrences that fulfill the predicate, discard the rest.

filterApply :: Behavior (a -> Bool) -> Event a -> Event a

Return all event occurrences that fulfill the time-varying predicate, discard the rest. Generalization of filterE.

whenE :: Behavior Bool -> Event a -> Event a

Return event occurrences only when the behavior is True. Variant of filterApply.

concatenate :: [a -> a] -> a -> a

Apply a list of functions in succession. Useful in conjunction with unions.

concatenate [f,g,h] = f . g . h

accumB :: MonadIO m => a -> Event (a -> a) -> m (Behavior a)

The accumB function is similar to a strict left fold, foldl'. It starts with an initial value and combines it with incoming events. For example, think

accumB "x" [(time1,(++"y")),(time2,(++"z"))]
   = stepper "x" [(time1,"xy"),(time2,"xyz")]

Note that the value of the behavior changes "slightly after" the events occur. This allows for recursive definitions.

unsafeMapIO :: (a -> IO b) -> Event a -> Event b

newEventsNamed

Arguments

:: Ord name

=> Handler (name, Event a, Handler a)
Initialization procedure.

-> IO (name -> Event a)
Series of events.

Create a series of events with delayed initialization.

For each name, the initialization handler will be called exactly once when the event is first "brought to life", e.g. when an event handler is registered to it.

test :: IO (Int -> IO ())

test_recursion1 :: IO (IO ())

Attributes

For a list of predefined attributes, see Graphics.UI.Threepenny.Attributes.

(#) :: a -> (a -> b) -> b infixl 8

Reverse function application. Allows convenient notation for setting properties.

Example usage.

mkElement "div"
    # set style     [("color","#CCAABB")]
    # set draggable True
    # set children  otherElements

(#.) :: UI Element -> String -> UI Element infixl 8

Convenient combinator for setting the CSS class on element creation.

type Attr x a = ReadWriteAttr x a a

Attributes can be set and get.

type WriteAttr x i = ReadWriteAttr x i ()

Attribute that only supports the set operation.

type ReadAttr x o = ReadWriteAttr x () o

Attribute that only supports the get operation.

data ReadWriteAttr x i o

Generalized attribute with different types for getting and setting.

Constructors

ReadWriteAttr

Fields

get' :: x -> UI o

set' :: i -> x -> UI ()

set :: ReadWriteAttr x i o -> i -> UI x -> UI x

Set value of an attribute in the UI monad. Best used in conjunction with #.

sink :: ReadWriteAttr x i o -> Behavior i -> UI x -> UI x

Set the value of an attribute to a Behavior, that is a time-varying value.

Note: For reasons of efficiency, the attribute is only updated when the value changes.

get :: ReadWriteAttr x i o -> x -> UI o

Get attribute value.

mkReadWriteAttr

Arguments

:: (x -> UI o)
Getter.

-> (i -> x -> UI ())
Setter.

-> ReadWriteAttr x i o

Build an attribute from a getter and a setter.

mkWriteAttr :: (i -> x -> UI ()) -> WriteAttr x i

Build attribute from a setter.

mkReadAttr :: (x -> UI o) -> ReadAttr x o

Build attribute from a getter.

bimapAttr :: (i' -> i) -> (o -> o') -> ReadWriteAttr x i o -> ReadWriteAttr x i' o'

Map input and output type of an attribute.

fromObjectProperty :: (FromJS a, ToJS a) => String -> Attr Element a

Turn a JavaScript object property .prop = ... into an attribute.

Widgets

class Widget w where

Widgets are data types that have a visual representation.

Methods

getElement :: w -> Element

element :: MonadIO m => Widget w => w -> m Element

Convenience synonym for return to make elements work well with set. Also works on Widgets.

Example usage.

e <- mkElement "button"
element e # set text "Ok"

widget :: Widget w => w -> UI w

Convenience synonym for return to make widgets work well with set.

JavaScript FFI

Direct interface to JavaScript in the browser window.

debug :: String -> UI ()

Print a message on the client console if the client has debugging enabled.

timestamp :: UI ()

Print a timestamp and the difference to the previous timestamp on the client console if the client has debugging enabled.

class ToJS a

Helper class for rendering Haskell values as JavaScript expressions.

Minimal complete definition

render

class FFI a

Helper class for making ffi a variable argument function.

Minimal complete definition

fancy

data JSFunction a

A JavaScript function with a given output type a.

ffi :: FFI a => String -> a

Simple JavaScript FFI with string substitution.

Inspired by the Fay language. https://github.com/faylang/fay/wiki

example :: String -> Int -> JSFunction String
example = ffi "$(%1).prop('checked',%2)"

The ffi function takes a string argument representing the JavaScript code to be executed on the client. Occurrences of the substrings %1 to %9 will be replaced by subequent arguments. The substring %% in the original will be replaced by % (character escape).

Note: Always specify a type signature! The types automate how values are marshalled between Haskell and JavaScript. The class instances for the FFI class show which conversions are supported.

runFunction :: JSFunction () -> UI ()

Run a JavaScript function, but do not wait for a result.

The client window uses JavaScript's eval() function to run the code.

NOTE: The JavaScript function need not be executed immediately, it can be buffered and sent to the browser window at a later time. See setCallBufferMode and flushCallBuffer for more.

callFunction :: JSFunction a -> UI a

Call a JavaScript function and wait for the result.

The client window uses JavaScript's eval() function to run the code.

data CallBufferMode

Specification of how JavaScript functions should be called.

Constructors

NoBuffering
When runFunction is used to call a JavaScript function, immediately send a message to the browser window to execute said function.

BufferRun
When runFunction is used to call a JavaScript function, hold back any message to the server. All JavaScript functions that are held back in this way are combined into a single message, which is finally sent whenever callFunction or flushCallBuffer are used, or an exported Haskell function is called.

FlushOften
The same as BufferRun, but this mode indicates client libraries and programs are encouraged to flush the buffer more often to simplify usage. Users may choose BufferRun instead if they want more control over flushing the buffer.

FlushPeriodically
The same as BufferRun, except that the buffer will also be flushed every 300ms.

setCallBufferMode :: CallBufferMode -> UI ()

Set the call buffering mode for the browser window.

flushCallBuffer :: UI ()

Flush the call buffer, i.e. send all outstanding JavaScript to the client in one single message.

ffiExport :: IsHandler a => a -> UI JSObject

Export the given Haskell function so that it can be called from JavaScript code.

NOTE: At the moment, the JSObject representing the exported function will be referenced by the browser Window in which it was created, preventing garbage collection until this browser Window is disconnected.

This makes it possible to use it as an event handler on the JavaScript side, but it also means that the Haskell runtime has no way to detect early when it is no longer needed.

In contrast, if you use the function domEvent to register an event handler to an Element, then the handler will be garbage collected as soon as the associated Element is garbage collected.

Internals

toJSObject :: Element -> JSObject

Access to the primitive JSObject for roll-your-own foreign calls.

liftJSWindow :: (Window -> IO a) -> UI a

Access to the primitive Window object, for roll-your-own JS foreign calls.

Internal and oddball functions

fromJQueryProp :: String -> (Value -> a) -> (a -> Value) -> Attr Element a

Turn a jQuery property .prop() into an attribute.