Graphics.UI.Threepenny.Elements

Description

Predefined DOM elements, for convenience.

Combinations and utilities

addStyleSheet :: Window -> FilePath -> UI ()

Add a stylesheet to the head.

The second argument refers to the filename of the stylesheet, but not its complete filepath. Threepenny will prefix the css subdirectory of the tpStatic configuration field to construct the complete filepath.

new :: UI Element

Make a new div element, synonym for div.

Primitive HTML elements

address :: UI Element

a :: UI Element

anchor :: UI Element

applet :: UI Element

area :: UI Element

audio :: UI Element

basefont :: UI Element

big :: UI Element

blockquote :: UI Element

body :: UI Element

bold :: UI Element

br :: UI Element

button :: UI Element

canvas :: UI Element

caption :: UI Element

center :: UI Element

cite :: UI Element

code :: UI Element

ddef :: UI Element

define :: UI Element

div :: UI Element

dlist :: UI Element

dterm :: UI Element

emphasize :: UI Element

fieldset :: UI Element

font :: UI Element

form :: UI Element

frame :: UI Element

frameset :: UI Element

h1 :: UI Element

h2 :: UI Element

h3 :: UI Element

h4 :: UI Element

h5 :: UI Element

h6 :: UI Element

header :: UI Element

hr :: UI Element

img :: UI Element

image :: UI Element

input :: UI Element

italics :: UI Element

keyboard :: UI Element

label :: UI Element

legend :: UI Element

li :: UI Element

link :: UI Element

map :: UI Element

meta :: UI Element

noframes :: UI Element

olist :: UI Element

option :: UI Element

p :: UI Element

paragraph :: UI Element

param :: UI Element

pre :: UI Element

sample :: UI Element

select :: UI Element

small :: UI Element

source :: UI Element

span :: UI Element

strong :: UI Element

sub :: UI Element

sup :: UI Element

table :: UI Element

td :: UI Element

textarea :: UI Element

th :: UI Element

thebase :: UI Element

thehtml :: UI Element

title_ :: UI Element

tr :: UI Element

tt :: UI Element

ul :: UI Element

underline :: UI Element

variable :: UI Element

video :: UI Element