Graphics.UI.Threepenny.SVG

Synopsis

SVG elements and attributes.

It is recommended that you import this module qualified, i.e.

import qualified Graphics.UI.Threepenny.SVG as SVG

Note that some SVG attributes and elements have the same name. In this case, only the most common of either has been imported. If you need the other version, you have to import the individual modules.

id :: WriteAttr Element String

mask :: WriteAttr Element String

filter :: WriteAttr Element String

max :: WriteAttr Element String

min :: WriteAttr Element String

rotate :: WriteAttr Element String

fill :: WriteAttr Element String

exponent :: WriteAttr Element Float

string :: WriteAttr Element String

from :: WriteAttr Element Float

to :: WriteAttr Element Float

local :: WriteAttr Element String

operator :: WriteAttr Element String

mode :: WriteAttr Element String

style :: WriteAttr Element String

clip :: WriteAttr Element String

version :: WriteAttr Element Float

by :: WriteAttr Element Float

class_ :: WriteAttr Element String

color :: WriteAttr Element String

height :: WriteAttr Element String

lang :: WriteAttr Element String

method :: WriteAttr Element String

name :: WriteAttr Element String

target :: WriteAttr Element String

type_ :: WriteAttr Element String

width :: WriteAttr Element String

stroke :: WriteAttr Element String

cursor :: WriteAttr Element String

accent_height :: WriteAttr Element Float

accumulate :: WriteAttr Element String

additive :: WriteAttr Element String

alphabetic :: WriteAttr Element Float

amplitude :: WriteAttr Element Float

arabic_form :: WriteAttr Element String

ascent :: WriteAttr Element Float

attributeName :: WriteAttr Element String

attributeType :: WriteAttr Element String

azimuth :: WriteAttr Element Float

baseFrequency :: WriteAttr Element String

baseProfile :: WriteAttr Element String

bbox :: WriteAttr Element String

begin :: WriteAttr Element String

bias :: WriteAttr Element Float

calcMode :: WriteAttr Element String

cap_height :: WriteAttr Element Float

clipPathUnits :: WriteAttr Element String

contentScriptType :: WriteAttr Element String

contentStyleType :: WriteAttr Element String

cx :: WriteAttr Element String

cy :: WriteAttr Element String

d :: WriteAttr Element String

descent :: WriteAttr Element Float

diffuseConstant :: WriteAttr Element Float

divisor :: WriteAttr Element Float

dur :: WriteAttr Element String

dx :: WriteAttr Element String

dy :: WriteAttr Element String

edgeMode :: WriteAttr Element String

elevation :: WriteAttr Element Float

end :: WriteAttr Element String

externalResourcesRequired :: WriteAttr Element String

filterRes :: WriteAttr Element String

filterUnits :: WriteAttr Element String

format :: WriteAttr Element String

fx :: WriteAttr Element String

fy :: WriteAttr Element String

g1 :: WriteAttr Element String

g2 :: WriteAttr Element String

glyph_name :: WriteAttr Element String

gradientTransform :: WriteAttr Element String

gradientUnits :: WriteAttr Element String

hanging :: WriteAttr Element Float

horiz_adv_x :: WriteAttr Element Float

horiz_origin_x :: WriteAttr Element Float

horiz_origin_y :: WriteAttr Element Float

ideographic :: WriteAttr Element Float

in_ :: WriteAttr Element String

in2 :: WriteAttr Element String

intercept :: WriteAttr Element Float

k :: WriteAttr Element Float

k1 :: WriteAttr Element Float

k2 :: WriteAttr Element Float

k3 :: WriteAttr Element Float

k4 :: WriteAttr Element Float

kernelMatrix :: WriteAttr Element String

kernelUnitLength :: WriteAttr Element String

keyPoints :: WriteAttr Element String

keySplines :: WriteAttr Element String

keyTimes :: WriteAttr Element String

lengthAdjust :: WriteAttr Element String

limitingConeAngle :: WriteAttr Element Float

markerHeight :: WriteAttr Element String

markerUnits :: WriteAttr Element String

markerWidth :: WriteAttr Element String

maskContentUnits :: WriteAttr Element String

maskUnits :: WriteAttr Element String

mathematical :: WriteAttr Element Float

media :: WriteAttr Element String

numOctaves :: WriteAttr Element Int

offset :: WriteAttr Element Float

onabort :: WriteAttr Element String

onactivate :: WriteAttr Element String

onbegin :: WriteAttr Element String

onclick :: WriteAttr Element String

onend :: WriteAttr Element String

onerror :: WriteAttr Element String

onfocusin :: WriteAttr Element String

onfocusout :: WriteAttr Element String

onload :: WriteAttr Element String

onmousedown :: WriteAttr Element String

onmousemove :: WriteAttr Element String

onmouseout :: WriteAttr Element String

onmouseover :: WriteAttr Element String

onmouseup :: WriteAttr Element String

onrepeat :: WriteAttr Element String

onresize :: WriteAttr Element String

onscroll :: WriteAttr Element String

onunload :: WriteAttr Element String

onzoom :: WriteAttr Element String

order :: WriteAttr Element String

orient :: WriteAttr Element String

orientation :: WriteAttr Element String

origin :: WriteAttr Element String

overline_position :: WriteAttr Element Float

overline_thickness :: WriteAttr Element Float

panose_1 :: WriteAttr Element Int

pathLength :: WriteAttr Element Float

patternContentUnits :: WriteAttr Element String

patternTransform :: WriteAttr Element String

patternUnits :: WriteAttr Element String

points :: WriteAttr Element String

pointsAtX :: WriteAttr Element Float

pointsAtY :: WriteAttr Element Float

pointsAtZ :: WriteAttr Element Float

preserveAlpha :: WriteAttr Element String

preserveAspectRatio :: WriteAttr Element String

primitiveUnits :: WriteAttr Element String

r :: WriteAttr Element String

radius :: WriteAttr Element String

refX :: WriteAttr Element String

refY :: WriteAttr Element String

rendering_intent :: WriteAttr Element String

repeatCount :: WriteAttr Element String

repeatDur :: WriteAttr Element String

requiredExtensions :: WriteAttr Element String

requiredFeatures :: WriteAttr Element String

restart :: WriteAttr Element String

result :: WriteAttr Element String

rx :: WriteAttr Element String

ry :: WriteAttr Element String

scale :: WriteAttr Element Float

seed :: WriteAttr Element Float

slope :: WriteAttr Element Float

spacing :: WriteAttr Element String

specularConstant :: WriteAttr Element Float

specularExponent :: WriteAttr Element Float

spreadMethod :: WriteAttr Element String

startOffset :: WriteAttr Element String

stdDeviation :: WriteAttr Element String

stemh :: WriteAttr Element Float

stemv :: WriteAttr Element Float

stitchTiles :: WriteAttr Element String

strikethrough_position :: WriteAttr Element Float

strikethrough_thickness :: WriteAttr Element Float

surfaceScale :: WriteAttr Element Float

systemLanguage :: WriteAttr Element String

tableValues :: WriteAttr Element String

targetX :: WriteAttr Element Float

targetY :: WriteAttr Element Float

textLength :: WriteAttr Element String

transform :: WriteAttr Element String

u1 :: WriteAttr Element String

u2 :: WriteAttr Element String

underline_position :: WriteAttr Element Float

underline_thickness :: WriteAttr Element Float

unicode :: WriteAttr Element String

unicode_range :: WriteAttr Element String

units_per_em :: WriteAttr Element Float

v_alphabetic :: WriteAttr Element Float

v_hanging :: WriteAttr Element Float

v_ideographic :: WriteAttr Element Float

v_mathematical :: WriteAttr Element Float

values :: WriteAttr Element String

vert_adv_y :: WriteAttr Element Float

vert_origin_x :: WriteAttr Element Float

vert_origin_y :: WriteAttr Element Float

viewBox :: WriteAttr Element String

viewTarget :: WriteAttr Element String

widths :: WriteAttr Element String

x :: WriteAttr Element String

x_height :: WriteAttr Element Float

x1 :: WriteAttr Element String

x2 :: WriteAttr Element String

xChannelSelector :: WriteAttr Element String

xlink_actuate :: WriteAttr Element String

xlink_arcrole :: WriteAttr Element String

xlink_href :: WriteAttr Element String

xlink_role :: WriteAttr Element String

xlink_show :: WriteAttr Element String

xlink_title :: WriteAttr Element String

xlink_type :: WriteAttr Element String

xml_base :: WriteAttr Element String

xml_lang :: WriteAttr Element String

xml_space :: WriteAttr Element String

y :: WriteAttr Element String

y1 :: WriteAttr Element String

y2 :: WriteAttr Element String

yChannelSelector :: WriteAttr Element String

z :: WriteAttr Element Float

zoomAndPan :: WriteAttr Element String

refx :: WriteAttr Element String

Deprecated: Use refX and refY instead

refy :: WriteAttr Element String

Deprecated: Use refX and refY instead

alignment_baseline :: WriteAttr Element String

baseline_shift :: WriteAttr Element String

clip_path :: WriteAttr Element String

clip_rule :: WriteAttr Element String

color_interpolation_filters :: WriteAttr Element String

color_interpolation :: WriteAttr Element String

color_profile :: WriteAttr Element String

color_rendering :: WriteAttr Element String

direction :: WriteAttr Element String

display :: WriteAttr Element String

dominant_baseline :: WriteAttr Element String

enable_background :: WriteAttr Element String

fill_opacity :: WriteAttr Element String

fill_rule :: WriteAttr Element String

flood_color :: WriteAttr Element String

flood_opacity :: WriteAttr Element String

font_family :: WriteAttr Element String

font_size_adjust :: WriteAttr Element String

font_size :: WriteAttr Element String

font_stretch :: WriteAttr Element String

font_style :: WriteAttr Element String

font_variant :: WriteAttr Element String

font_weight :: WriteAttr Element String

glyph_orientation_horizontal :: WriteAttr Element String

glyph_orientation_vertical :: WriteAttr Element String

image_rendering :: WriteAttr Element String

kerning :: WriteAttr Element String

letter_spacing :: WriteAttr Element String

lighting_color :: WriteAttr Element String

marker_end :: WriteAttr Element String

marker_mid :: WriteAttr Element String

marker_start :: WriteAttr Element String

opacity :: WriteAttr Element String

overflow :: WriteAttr Element String

pointer_events :: WriteAttr Element String

shape_rendering :: WriteAttr Element String

stop_color :: WriteAttr Element String

stop_opacity :: WriteAttr Element String

stroke_dasharray :: WriteAttr Element String

stroke_dashoffset :: WriteAttr Element String

stroke_linecap :: WriteAttr Element String

stroke_linejoin :: WriteAttr Element String

stroke_miterlimit :: WriteAttr Element String

stroke_opacity :: WriteAttr Element String

stroke_width :: WriteAttr Element String

text_anchor :: WriteAttr Element String

text_decoration :: WriteAttr Element String

text_rendering :: WriteAttr Element String

unicode_bidi :: WriteAttr Element String

visibility :: WriteAttr Element String

word_spacing :: WriteAttr Element String

writing_mode :: WriteAttr Element String

text :: UI Element

path :: UI Element

symbol :: UI Element

pattern :: UI Element

linearGradient :: UI Element

title :: UI Element

set :: UI Element

a :: UI Element

font :: UI Element

image :: UI Element

glyphRef :: UI Element

altGlyph :: UI Element

altGlyphDef :: UI Element

altGlyphItem :: UI Element

animate :: UI Element

animateColor :: UI Element

animateMotion :: UI Element

animateTransform :: UI Element

circle :: UI Element

clipPath :: UI Element

colorProfile :: UI Element

defs :: UI Element

desc :: UI Element

ellipse :: UI Element

feBlend :: UI Element

feColorMatrix :: UI Element

feComponentTransfer :: UI Element

feComposite :: UI Element

feConvolveMatrix :: UI Element

feDiffuseLighting :: UI Element

feDisplacementMap :: UI Element

feDistantLight :: UI Element

feFlood :: UI Element

feFuncA :: UI Element

feFuncB :: UI Element

feFuncG :: UI Element

feFuncR :: UI Element

feGaussianBlur :: UI Element

feImage :: UI Element

feMerge :: UI Element

feMergeNode :: UI Element

feMorphology :: UI Element

feOffset :: UI Element

fePointLight :: UI Element

feSpecularLighting :: UI Element

feSpotLight :: UI Element

feTile :: UI Element

feTurbulence :: UI Element

font_face :: UI Element

font_face_format :: UI Element

font_face_name :: UI Element

font_face_srv :: UI Element

font_face_uri :: UI Element

foreignObject :: UI Element

g :: UI Element

glyph :: UI Element

hkern :: UI Element

line :: UI Element

marker :: UI Element

metadata :: UI Element

missing_glyph :: UI Element

mpath :: UI Element

polygon :: UI Element

polyline :: UI Element

radialGradient :: UI Element

rect :: UI Element

script :: UI Element

stop :: UI Element

svg :: UI Element

switch :: UI Element

textPath :: UI Element

tref :: UI Element

tspan :: UI Element

use :: UI Element

view :: UI Element

vkern :: UI Element