Graphics.UI.Threepenny.Timer

Synopsis

Implementation of a simple timer which runs on the server-side.

NOTE: The timer may be rather wobbly unless you compile with the -threaded option.

Documentation

data Timer

timer :: UI Timer

Create a new timer

interval :: Attr Timer Int

Timer interval in milliseconds.

running :: Attr Timer Bool

Whether the timer is running or not.

tick :: Timer -> Event ()

Timer event.

start :: Timer -> UI ()

Start the timer.

stop :: Timer -> UI ()

Stop the timer.