Graphics.UI.Threepenny.Widgets

Synopsis

Widgets are reusable building blocks for a graphical user interface. This module collects useful widgets that are designed to work with functional reactive programming (FRP).

For more details and information on how to write your own widgets, see the widget design guide.

Tidings

data Tidings a

Data type representing a behavior (facts) and suggestions to change it (rumors).

rumors :: Tidings a -> Event a

facts :: Tidings a -> Behavior a

tidings :: Behavior a -> Event a -> Tidings a

Smart constructor. Combine facts and rumors into Tidings.

Widgets

Input widgets

data TextEntry

A single-line text entry.

entry

Arguments

:: Behavior String
Display value when the element does not have focus.

-> UI TextEntry

Create a single-line text entry.

userText :: TextEntry -> Tidings String

User changes to the text value.

ListBox

data ListBox a

A list of values. The user can select entries.

listBox

Arguments

:: forall a. Ord a

=> Behavior [a]
list of items

-> Behavior (Maybe a)
selected item

-> Behavior (a -> UI Element)
display for an item

-> UI (ListBox a)

Create a ListBox.

userSelection :: ListBox a -> Tidings (Maybe a)

User changes to the current selection (possibly empty).